Move db item expansion models to db-item-expansion
